fetch-users/requirements.txt
pydantic==1.10.7
python-dotenv==1.0.0
requests==2.28.2
pymongo==4.3.3